app/jurisbot/legal-store-search/requirements.txt
dash
dash-bootstrap-components
requests
google-auth
gunicorn
python-dotenv
boto3
PyYAML